requirements.txt
streamlit
python-dotenv
pymupdf
faiss-cpu
langchain
langchain-community
langchain-openai
tiktoken
